feature: changed folder naming
