--- mp-extended/src/main/java/com/mercadopago/sdk/android/coremethods/builder/data/datasource/remote/mapper/DeviceSessionResponseMapper.kt
@@ -1,9 +1,0 @@
-package com.mercadopago.sdk.android.coremethods.builder.data.datasource.remote.mapper
-
-import com.mercadopago.sdk.android.coremethods.builder.data.remote.response.DeviceSessionIdResponse
-import com.mercadopago.sdk.android.coremethods.builder.domain.model.MPDeviceSession
-
-internal fun DeviceSessionIdResponse.toModel() =
-    MPDeviceSession(
-        session = this.sessionId,
-    )
