--- mp-extended/src/main/java/com/mercadopago/sdk/android/coremethods/builder/data/remote/mappers/GetDeviceSessionIdParamsMapper.kt
@@ -1,7 +1,0 @@
-package com.mercadopago.sdk.android.coremethods.builder.data.remote.mappers
-
-import com.mercadopago.sdk.android.coremethods.builder.data.remote.request.MPDeviceSessionIdRequest
-import com.mercadopago.sdk.android.coremethods.builder.domain.model.params.GetDeviceSessionParams
-
-internal fun GetDeviceSessionParams.toRequest(): MPDeviceSessionIdRequest =
-    MPDeviceSessionIdRequest(fingerprint = this.device?.fingerprint, siteId = this.siteId)
