+++ mp-extended/src/main/java/com/mercadopago/sdk/android/mpextended/di/repository/RepositoryModule.kt
@@ -1,0 +1,13 @@
+package com.mercadopago.sdk.android.mpextended.di.repository
+
+import com.mercadopago.sdk.android.mpextended.data.repository.MpExtendedRepositoryImpl
+import com.mercadopago.sdk.android.mpextended.domain.repository.MPExtendedRepository
+import org.koin.core.module.Module
+import org.koin.dsl.module
+
+internal fun provideRepositoryModule(): Module =
+    module {
+        factory<MPExtendedRepository> {
+            MpExtendedRepositoryImpl(get())
+        }
+    }
